fetch/fetch/requirements.txt
Django==4.0.7
django-crispy-forms==1.13.0
django-filter==21.1
django-import-export==2.7.1
django-registration-redux==2.9
django-tinymce==3.4.0
django-versatileimagefield==2.2
psycopg2-binary==2.9.3
python-decouple==3.5
python-magic==0.4.24
pytz==2021.3